Rename DataOverviewFragment to DatasetListFragment
To be more accurate about what the fragment actually is. Also move the
files corresponding to the DatasetListFragment into view/fragments.
